精简联系表单并移除代理门面
